Adding support for multiple listeners and renaming.
